Move EthBridge VP into vp.rs
