+++ apps/src/lib/wallet/defaults.rs
@@ -21,7 +21,7 @@ pub fn addresses_from_genesis(genesis: GenesisConfig) -> Vec<(Alias, Address)> {
         ("pos".into(), pos::ADDRESS),
         ("pos_slash_pool".into(), pos::SLASH_POOL_ADDRESS),
         ("governance".into(), governance::vp::ADDRESS),
-        ("eth_bridge".into(), eth_bridge::ADDRESS),
+        ("eth_bridge".into(), eth_bridge::vp::ADDRESS),
     ];
     // Genesis validators
     let validator_addresses =
